fix last merge command test
