+++ src/lib/src/repositories/checkout.rs
@@ -59,7 +59,7 @@ pub fn checkout_theirs(repo: &LocalRepository, path: impl AsRef<Path>) -> Result
         .find(|c| c.merge_entry.path == path.as_ref())
     {
         // Lookup the file for the merge commit entry and copy it over
-        command::restore(
+        repositories::restore::restore(
             repo,
             RestoreOpts::from_path_ref(path, conflict.merge_entry.commit_id.clone()),
         )
@@ -114,13 +114,25 @@ pub fn checkout_combine<P: AsRef<Path>>(repo: &LocalRepository, path: P) -> Resu
                 &conflict.base_entry.hash,
                 &conflict.base_entry.filename,
             );
-            let df_base = tabular::read_df(df_base_path, DFOpts::empty())?;
+            let df_base = tabular::maybe_read_df_with_extension(
+                repo,
+                &df_base_path,
+                &conflict.base_entry.path,
+                &conflict.base_entry.commit_id,
+                &DFOpts::empty(),
+            )?;
             let df_merge_path = util::fs::version_path_from_hash_and_filename(
                 repo,
                 &conflict.merge_entry.hash,
                 &conflict.merge_entry.filename,
             );
-            let df_merge = tabular::read_df(df_merge_path, DFOpts::empty())?;
+            let df_merge = tabular::maybe_read_df_with_extension(
+                repo,
+                df_merge_path,
+                &conflict.merge_entry.path,
+                &conflict.merge_entry.commit_id,
+                &DFOpts::empty(),
+            )?;
 
             log::debug!("GOT DF HEAD {}", df_base);
             log::debug!("GOT DF MERGE {}", df_merge);
